Simplify and centralize instance state determination

+++ stratosphere/submodels/compute_group.py
@@ -27,15 +27,9 @@ class InstanceStatesSnapshot(models.Model):
     user_configuration = models.ForeignKey('UserConfiguration', related_name='instance_states_snapshots')
     time = models.DateTimeField()
 
-    running    = models.IntegerField()
-    rebooting  = models.IntegerField()
-    terminated = models.IntegerField()
     pending    = models.IntegerField()
-    stopped    = models.IntegerField()
-    suspended  = models.IntegerField()
-    paused     = models.IntegerField()
-    error      = models.IntegerField()
-    unknown    = models.IntegerField()
+    running    = models.IntegerField()
+    terminated = models.IntegerField()
 
 
 class GroupInstanceStatesSnapshot(models.Model):
@@ -45,15 +39,9 @@ class GroupInstanceStatesSnapshot(models.Model):
     user_snapshot = models.ForeignKey('InstanceStatesSnapshot', related_name='group_snapshots')
     group = models.ForeignKey('ComputeGroup', related_name='instance_states_snapshots')
 
-    running    = models.IntegerField()
-    rebooting  = models.IntegerField()
-    terminated = models.IntegerField()
     pending    = models.IntegerField()
-    stopped    = models.IntegerField()
-    suspended  = models.IntegerField()
-    paused     = models.IntegerField()
-    error      = models.IntegerField()
-    unknown    = models.IntegerField()
+    running    = models.IntegerField()
+    terminated = models.IntegerField()
 
 
 class ComputeGroupBase(models.Model, HasLogger, SaveTheChange):
@@ -62,13 +50,11 @@ class ComputeGroupBase(models.Model, HasLogger, SaveTheChange):
 
     PENDING = 'PENDING'
     RUNNING = 'RUNNING'
-    STOPPED = 'STOPPED'
     TERMINATED = 'TERMINATED'
 
     STATE_CHOICES = (
         (PENDING, 'Pending'),
         (RUNNING, 'Running'),
-        (STOPPED, 'Stopped'),
         (TERMINATED, 'Terminated'),
     )
 
@@ -82,32 +68,29 @@ class ComputeGroupBase(models.Model, HasLogger, SaveTheChange):
     size_distribution = JSONField()
     state = models.CharField(max_length=16, choices=STATE_CHOICES, default=PENDING)
     authentication_method = models.ForeignKey('AuthenticationMethod', related_name='compute_groups')
-    last_state_update_time = models.DateTimeField()
+
+    @classmethod
+    def quick_create(cls, instance_count=5):
+        from ..models import UserConfiguration, OperatingSystemImage, AuthenticationMethod, Provider
+        image = OperatingSystemImage.objects.first()
+        providers = Provider.objects.filter(disk_image_mappings__operating_system_image=image)
+        provider_policy = {p.name: 'auto' for p in providers}
+        return cls.objects.create(user_configuration=UserConfiguration.objects.first(),
+                                  image=image, cpu=1, memory=512,
+                                  instance_count=instance_count, name='sdfsdf', provider_policy=provider_policy,
+                                  size_distribution={}, authentication_method=AuthenticationMethod.objects.first())
 
     def provider_states(self):
         provider_states_map = {}
         now = timezone.now()
-        two_minutes_ago = now - timedelta(minutes=2)
-        four_minutes_ago = now - timedelta(minutes=4)
-
-        def instance_terminated(i):
-            pending_or_running_states = None, ComputeInstance.RUNNING, ComputeInstance.PENDING
-            not_pending_or_running = i.state not in pending_or_running_states
-            pending_or_running_and_expired = i.state in pending_or_running_states and i.last_request_start_time < two_minutes_ago
-            unknown_and_not_expired = i.state not in pending_or_running_states and i.last_state_update_time >= four_minutes_ago
-            return not_pending_or_running or pending_or_running_and_expired or unknown_and_not_expired
 
         for provider_name in self.provider_policy:
             provider_configuration = ProviderConfiguration.objects.get(provider_name=provider_name, user_configuration=self.user_configuration)
             provider_instances = self.instances.filter(provider_image__provider__name=provider_name, terminated=False)
 
-            running_count = len(list(filter(lambda i: i.state == ComputeInstance.RUNNING, provider_instances)))
-            pending_count = len(list(filter(lambda i: i.state in (None, ComputeInstance.PENDING) and i.last_request_start_time >= two_minutes_ago, provider_instances)))
-
-            if running_count >= self.instance_count:
-                terminated_count = 0
-            else:
-                terminated_count = len(list(filter(instance_terminated, provider_instances)))
+            running_count = len(list(filter(lambda i: i.is_running(now), provider_instances)))
+            pending_count = len(list(filter(lambda i: i.is_pending(now), provider_instances)))
+            terminated_count = len(list(filter(lambda i: i.is_terminated(now), provider_instances)))
 
             icon_path = staticfiles_storage.url(provider_configuration.provider.icon_path)
 
@@ -124,54 +107,52 @@ class ComputeGroupBase(models.Model, HasLogger, SaveTheChange):
     @thread_local(DB_OVERRIDE='serializable')
     def check_instance_distribution(self):
         with transaction.atomic():
-            instances_flat = list(self.instances.all())
-
-            running_count = len(list(filter(lambda i: i.state == ComputeInstance.RUNNING, instances_flat)))
-            non_pending_count = len(list(filter(lambda i: i.state not in (None, ComputeInstance.PENDING), instances_flat)))
-            non_terminated_count = len(list(filter(lambda i: i.state not in (ComputeInstance.TERMINATED, ComputeInstance.UNKNOWN), instances_flat)))
+            now = timezone.now()
 
             if self.state == self.TERMINATED:
+                non_terminated_count = self.instances.filter(~ComputeInstance.terminated_instances_query(now)).count()
                 self.logger.info('Compute group state is TERMINATED. Remaining non-terminated instances: %d' % non_terminated_count)
                 if non_terminated_count == 0:
                     self.logger.warn('Deleting self')
                     self.delete()
 
-            elif self.state == self.PENDING and non_pending_count >= self.instance_count:
-                self.logger.info('State is currently pending and non-pending count >= instance_count (%d >= %d); ' + \
-                                 'setting group state to RUNNING' % (non_pending_count, self.instance_count))
-                self.state = self.RUNNING
-                self.save()
+            else:
+                self.logger.info('Group state is %s; ensuring running instance count no less than expected' % self.state)
 
-            elif self.state == self.RUNNING:
-                self.logger.info('Group state is RUNNING; ensuring running instance count no less than expected')
+                running_count = self.instances.filter(ComputeInstance.running_instances_query(now)).count()
                 self.logger.info('running_count < self.instance_count = %d < %d' % (running_count, self.instance_count))
 
                 if running_count < self.instance_count:
                     self.logger.warn('Running count is less than expected count')
 
                     # state__in=[None] returns empty list no matter what
-                    bad_instances = self.instances.exclude(state__in=(ComputeInstance.PENDING, ComputeInstance.RUNNING)).exclude(state=None)
+                    bad_instances = self.instances.filter(ComputeInstance.terminated_instances_query(now))
                     self.logger.warn('bad instances: %s' % [i.pk for i in bad_instances])
 
                     bad_provider_ids = set([instance.provider_configuration.pk for instance in bad_instances])
-                    good_provider_ids = [provider.pk for provider in
-                                         self.user_configuration.provider_configurations.exclude(pk__in=bad_provider_ids)]
-
-                    self.logger.warn('bad providers: %s' % bad_provider_ids)
-                    self.logger.warn('good providers: %s' % good_provider_ids)
 
                 else:
-                    good_provider_ids = self.user_configuration.provider_configurations.values_list('id', flat=True)
+                    bad_provider_ids = set()
 
-                if running_count != self.instance_count:
-                    self.logger.warn('Rebalancing instances')
-                    self.rebalance_instances(good_provider_ids)
+                if running_count >= self.instance_count and self.state == self.PENDING:
+                    self.state = self.RUNNING
+                    self.save()
+
+                good_provider_ids = [provider.pk for provider in
+                                     self.user_configuration.provider_configurations.exclude(pk__in=bad_provider_ids)]
+
+                self.logger.warn('bad providers: %s' % bad_provider_ids)
+                self.logger.warn('good providers: %s' % good_provider_ids)
+
+            if running_count != self.instance_count:
+                self.logger.warn('Rebalancing instances')
+                self.rebalance_instances(good_provider_ids)
 
     @thread_local(DB_OVERRIDE='serializable')
     def rebalance_instances(self, provider_ids=None):
         best_sizes = self._get_best_sizes(provider_ids)
 
-        # if there aren't any providers left, then fuck it; deploy the required number of
+        # if there aren't any providers left deploy the required number of
         # instances to every provider in hopes that one of them will work
         if len(best_sizes) == 0:
             self.logger.warn('No sizes available for any provider; rebalancing across all providers as a Hail Mary')
@@ -180,11 +161,13 @@ class ComputeGroupBase(models.Model, HasLogger, SaveTheChange):
         else:
             instance_counts_by_size_id = self._get_size_distribution(best_sizes)
 
-        self.logger.warn('New size distribution: %s' % instance_counts_by_size_id)
+        self.logger.info('New size distribution: %s' % instance_counts_by_size_id)
         self.size_distribution = instance_counts_by_size_id
         self.save()
 
         self._create_compute_instances()
+
+        self.user_configuration.take_instance_states_snapshot_if_changed()
 
     @thread_local(DB_OVERRIDE='serializable')
     def terminate_instance(self, instance):
@@ -195,6 +178,7 @@ class ComputeGroupBase(models.Model, HasLogger, SaveTheChange):
             instance.terminate()
 
     @retry(OperationalError)
+    @thread_local(DB_OVERRIDE='serializable')
     def terminate(self):
         with transaction.atomic():
             self.state = self.TERMINATED
@@ -205,21 +189,11 @@ class ComputeGroupBase(models.Model, HasLogger, SaveTheChange):
         # TODO figure out how to make this more consistent without causing a bunch of transaction conflicts
         instances = list(self.instances.all())
 
-        two_minutes_ago = now - timedelta(minutes=2)
-        def filter_instance_states(two_minutes_ago, states):
-            return lambda i: i.state in states and (not i.terminated or i.last_request_start_time < two_minutes_ago)
-
         args = {
             'group': self,
-            'running': len(list(filter(filter_instance_states(two_minutes_ago, [ComputeInstance.RUNNING]), instances))),
-            'rebooting': len(list(filter(filter_instance_states(two_minutes_ago, [ComputeInstance.REBOOTING]), instances))),
-            'terminated': len(list(filter(filter_instance_states(two_minutes_ago, [ComputeInstance.TERMINATED]), instances))),
-            'pending': len(list(filter(filter_instance_states(two_minutes_ago, [None, ComputeInstance.PENDING]), instances))),
-            'stopped': len(list(filter(filter_instance_states(two_minutes_ago, [ComputeInstance.STOPPED]), instances))),
-            'suspended': len(list(filter(filter_instance_states(two_minutes_ago, [ComputeInstance.SUSPENDED]), instances))),
-            'paused': len(list(filter(filter_instance_states(two_minutes_ago, [ComputeInstance.PAUSED]), instances))),
-            'error': len(list(filter(filter_instance_states(two_minutes_ago, [ComputeInstance.ERROR]), instances))),
-            'unknown': len(list(filter(filter_instance_states(two_minutes_ago, [ComputeInstance.UNKNOWN]), instances))),
+            'pending': len(list(filter(lambda i: i.is_pending(now), instances))),
+            'running': len(list(filter(lambda i: i.is_running(now), instances))),
+            'terminated': len(list(filter(lambda i: i.is_terminated(now), instances))),
         }
 
         return GroupInstanceStatesSnapshot(**args)
@@ -303,14 +277,10 @@ class ComputeGroupBase(models.Model, HasLogger, SaveTheChange):
 
     def _pending_or_running_count(self, provider_size):
         now = timezone.now()
-        return self.instances.filter(
-            Q(provider_size=provider_size)
-            & (Q(state=ComputeInstance.RUNNING)
-               | Q(state=None,
-                   last_request_start_time__gt=now - timedelta(minutes=2))
-               | Q(state=ComputeInstance.PENDING,
-                   last_request_start_time__gt=now - timedelta(minutes=5))
-        )).count()
+        query = Q(state__in=(ComputeInstance.PENDING, ComputeInstance.RUNNING)) | Q(state=None)
+        query &= Q(provider_size=provider_size)
+        instances = self.instances.filter(query)
+        return len([i.is_pending(now) or i.is_running(now) for i in instances])
 
     def _create_compute_instances(self):
         with transaction.atomic():
@@ -321,13 +291,17 @@ class ComputeGroupBase(models.Model, HasLogger, SaveTheChange):
                 provider_size = ProviderSize.objects.get(pk=provider_size_id)
                 provider_configuration = provider_size.provider_configuration
 
+                self.logger.info('Balancing compute instances for size %s; expected count %d' %
+                                 (provider_size, size_instance_count))
+
                 pending_or_running_count = self._pending_or_running_count(provider_size)
 
                 running_provider_instances = self.instances.filter(provider_size=provider_size,
                                                                    state=ComputeInstance.RUNNING)
                 running_count = running_provider_instances.count()
 
-                self.logger.info('pending_or_running_count: %d, running_count: %d' % (pending_or_running_count, running_count))
+                self.logger.info('pending_or_running_count: %d, running_count: %d' %
+                                 (pending_or_running_count, running_count))
 
                 # filter on provider as well, since available_provider_images could contain shared images
                 # TODO wait, does that make sense?
@@ -353,22 +327,31 @@ class ComputeGroupBase(models.Model, HasLogger, SaveTheChange):
                             'state': None,
                             'public_ips': [],
                             'private_ips': [],
-                            'last_request_start_time': timezone.now(),
                         }
 
                         compute_instance = ComputeInstance.objects.create(**compute_instance_args)
-                        print('created instance %d for provider_size %s' % (compute_instance.pk, provider_size))
-
-                        # lexical scoping in loop
-                        connection.on_commit(lambda c=compute_instance: c.schedule_create_libcloud_node_job())
+                        self.logger.info('Created instance %d for provider_size %s' % (compute_instance.pk, provider_size))
 
                 elif running_count > size_instance_count:
-                    terminate_instances.extend(list(running_provider_instances)[:running_count - size_instance_count])
+                    extra_instances = list(running_provider_instances)[:running_count - size_instance_count]
+                    self.logger.info('Found %d extra instances for size %s' % (len(extra_instances), provider_size))
+                    terminate_instances.extend(extra_instances)
 
             # TODO make sure this squares with multi-row transaction isolation
             missing_size_instances = self.instances.filter(~Q(provider_size__pk__in=self.size_distribution.keys()))
             terminate_instances.extend(missing_size_instances)
 
-            for instance in terminate_instances:
-                self.logger.warning('terminating instance %s for size %s' % (instance, instance.provider_size))
-                instance.terminate()
+            self.logger.info('Found %d instances for sizes not in size distribution' % len(missing_size_instances))
+
+            if len(terminate_instances) > 0:
+                total_running_count = self.instances.filter(state=ComputeInstance.RUNNING).count()
+                if total_running_count < self.instance_count:
+                    # Two motivations here: we stop ourselves from getting unlucky by accidentally terminating an
+                    # instance that's just having a bad minute, and we also stop ourselves from terminating instances
+                    # that are taking so long to start up that they're considered as failed. The latter can get into
+                    # a loop of continually starting, timing out, and restarting instances, otherwise.
+                    self.logger.warn('Not terminating extraneous instances while running count is less than or equal to expected count')
+                else:
+                    for instance in terminate_instances:
+                        self.logger.warn('Terminating instance %s for size %s' % (instance, instance.provider_size))
+                        instance.terminate()
